Create POC to merge both the bundle and app chart into the same repo
This is Proof of Concept to have both the App and crossplane resources in the same repository, and simplify installation.

The main idea is that there's an App chart, which is installed on the MC and installs the necessary Crossplane resources and the WC chart.

Atm the WC chart is installed as an App CR, but could easily be refactored into a HelmRelease.

To simplify installation, all the values intended for the actual WC chart are provided as is to the App chart, and will be passed over to the ConfigMap for the WC chart. This implies that the `values.yaml` and `values.schema.json` files are duplicated in both charts. Maybe we could keep just the ones from the App (parent) chart, or we could automate the sync between the two. Open to ideas.

Another question mark is about the release and catalog pushing automation, not sure how that would look like with two charts in the same repo.

Regarding versioning, I think we should keep the version of both charts linked together, and release them together as a single release/package.

Towards https://github.com/giantswarm/giantswarm/issues/34551